feat(auth): migrate next-auth v4 to Auth.js v5
Major auth infrastructure migration per Auth.js v5 guidelines:

- Create central auth.ts config file exporting handlers, auth, signIn, signOut
- Simplify app/api/auth/[...nextauth]/route.ts to re-export handlers
- Replace getServerSession() with auth() in lib/rbac.ts and lib/session-cache.ts
- Update lib/auth-depth.ts for v5 cookie names (authjs.*) with v4 fallback
- Support AUTH_SECRET env var (v5 standard) with NEXTAUTH_SECRET fallback
- Update type declarations for v5 module augmentation pattern
- Update test infrastructure: __mocks__/auth.ts bridges to next-auth mock
  for backward compatibility with 30+ existing test files
- Upgrade next-auth from ^4.24.11 to 5.0.0-beta.30

All 1002 existing passing tests continue to pass. The 21 pre-existing
failures (tracked in #276) are unrelated to this migration.

Fixes #200
