+++ .env.example
@@ -31,9 +31,10 @@ REDIS_PORT=6379
 # 格式：redis://:password@host:port
 REDIS_URL=redis://:changeme_redis_password_here@titan-redis:6379
 
-# ── TITAN Next.js App（Issue #185）──────────────────────────
-# 必填：NextAuth JWT 簽名密鑰（使用 openssl rand -hex 32 產生）
-NEXTAUTH_SECRET=changeme_nextauth_secret_here
+# ── TITAN Next.js App（Issue #185, #200）─────────────────────
+# 必填：Auth.js v5 JWT 簽名密鑰（使用 openssl rand -hex 32 產生）
+# Auth.js v5 使用 AUTH_SECRET（向下相容 NEXTAUTH_SECRET）
+AUTH_SECRET=changeme_auth_secret_here
 # TITAN 對外 URL（Nginx 反向代理後的完整 URL）
 TITAN_URL=https://titan.bank.local/titan
