+++ __mocks__/next-auth.ts
@@ -1,4 +1,10 @@
-// Mock next-auth for tests
+/**
+ * Mock next-auth for tests — supports both v4 and v5 patterns.
+ *
+ * Auth.js v5 migration (Issue #200):
+ * - getServerSession is kept for backward compatibility with existing tests
+ * - The @/auth mock (__mocks__/auth.ts) delegates to this getServerSession
+ */
 const mockSession = {
   user: {
     id: "user-1",
@@ -11,6 +17,7 @@ const mockSession = {
 
 export const getServerSession = jest.fn(() => Promise.resolve(mockSession));
 
+// Client-side exports (next-auth/react)
 export const useSession = jest.fn(() => ({
   data: mockSession,
   status: "authenticated",
