+++ package.json
@@ -31,7 +31,7 @@
     "ioredis": "^5.10.1",
     "lucide-react": "^0.469.0",
     "next": "^15.5.14",
-    "next-auth": "^4.24.11",
+    "next-auth": "^5.0.0-beta.30",
     "pino": "^10.3.1",
     "rate-limiter-flexible": "^10.0.1",
     "react": "^19.0.0",
